+++ src/devices/navigation2DClient/Navigation2DClient.cpp
@@ -579,7 +579,7 @@ bool yarp::dev::Navigation2DClient::getLocationsList(std::vector<std::string>& l
             if (list)
             {
                 locations.clear();
-                for (int i = 0; i < list->size(); i++)
+                for (size_t i = 0; i < list->size(); i++)
                 {
                     locations.push_back(list->get(i).asString());
                 }
